docker/web/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
conan==2.7.0
distro==1.8.0
Django==5.1.1
django-environ==0.11.2
django_components==0.95
fasteners==0.19
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
patch-ng==1.18.0
protobuf==5.27.0
psutil==6.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
urllib3==1.26.20
uWSGI==2.0.26
wheel==0.44.0
